+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
-- No pending changes
+### Changed
+- Deprecated smart getters and setters
 
 ## [0.7.0] - 2018-02-03
 ### Added